move methods from weblink to WebUtil to reduce dependency on Weblink class
